+++ lib/plotuty.py
@@ -1,3 +1,7 @@
+#*****************************************************************************************
+# module used to save a plot and delete old plots 
+#*****************************************************************************************
+
 __author__ = 'IU1BOW - Corrado'
 
 import logging
@@ -7,14 +11,15 @@ import glob
 import os
 import json
 
-logging.config.fileConfig("../cfg/log_config.ini", disable_existing_loggers=False)
+logging.config.fileConfig("../cfg/plots_log_config.ini", disable_existing_loggers=False)
+logger = logging.getLogger(__name__)
 
 def del_with_retention(filename, number,ext):
 #list all file with date time and extension
 #and delete al file except the last "number" passed by param, and ordered by name 
-    logging.info('deleting old files')
+    logger.info('deleting old files')
     pattern=filename+'_????????-??????.'+ext
-    logging.debug('search pattern for delete files: '+pattern)
+    logger.debug('search pattern for delete files: '+pattern)
     files=glob.glob(pattern)
     length = len(files)
     if length>number:
@@ -24,10 +29,10 @@ def del_with_retention(filename, number,ext):
         while i < length - number: 
             try:
                 os.remove(files_ordered[i])
-                logging.info('deleted file:'+files_ordered[i])
+                logger.info('deleted file:'+files_ordered[i])
             except Exception as e:
-                logging.error(e)
-                logging.error('not deleted file:'+files_ordered[i])
+                logger.error(e)
+                logger.error('not deleted file:'+files_ordered[i])
             finally:
                 i += 1
 
@@ -43,29 +48,29 @@ def saveplt(plt,filename):
     #saving svg file
     try:
         plt.savefig(outputfile+'.svg', dpi=100, bbox_inches='tight')    
-        logging.info('plotted saved on: '+outputfile+'.svg')
+        logger.info('plotted saved on: '+outputfile+'.svg')
     except Exception as e01:
-        logging.error(e01)
-        logging.error('error saving file: '+outputfile+'.svg')
+        logger.error(e01)
+        logger.error('error saving file: '+outputfile+'.svg')
 
     #saving png file
     try:
         plt.savefig(outputfile+'.png', dpi=100, bbox_inches='tight')  
-        logging.info('plotted saved on: '+outputfile+'.png')    
+        logger.info('plotted saved on: '+outputfile+'.png')    
     except Exception as e02:
-        logging.error(e02)
-        logging.error('error saving file: '+outputfile+'.png')
+        logger.error(e02)
+        logger.error('error saving file: '+outputfile+'.png')
 
     #creating idx file in order to match standard file name to filename with timestamp
     try:
         basepath=os.path.dirname(filename)
         idxfile=os.path.join(basepath,'plots.json')
         if os.path.exists(idxfile):
-            logging.debug(idxfile+' found')
+            logger.debug(idxfile+' found')
             with open(idxfile,'r') as jsonfile:
                 json_content = json.load(jsonfile) # this is now in memory! you can use it outside 'open'
         else:
-            logging.debug(idxfile+ ' not found')
+            logger.debug(idxfile+ ' not found')
             json_content={}
 
         json_content[os.path.basename(filename)] = os.path.basename(outputfile)
@@ -75,8 +80,8 @@ def saveplt(plt,filename):
         logging.info(idxfile+' updated')
 
     except Exception as e03:
-        logging.error(e03)
-        logging.error('error writing idx: '+idxfile)
+        logger.error(e03)
+        logger.error('error writing idx: '+idxfile)
 
     del_with_retention(filename,2,'svg')
     del_with_retention(filename,2,'png')