+++ scripts/config.sh
@@ -1,4 +1,4 @@
 DIR=`realpath -s $0|sed 's|\(.*\)/.*|\1|'` 
 echo Absolute path: ${DIR}  
 cd ${DIR}  
-python3 config.py
+python3 ../lib/config.py