+++ scripts/propagation_heatmaps.sh
@@ -1,4 +1,4 @@
 DIR=`realpath -s $0|sed 's|\(.*\)/.*|\1|'` 
 echo Absolute path: ${DIR}
 cd ${DIR} 
-python3 propagation_heatmaps.py
+python3 ../lib/propagation_heatmaps.py