+++ scripts/qso_hour_band.sh
@@ -1,4 +1,4 @@
 DIR=`realpath -s $0|sed 's|\(.*\)/.*|\1|'` 
 echo Absolute path: ${DIR}
 cd ${DIR} 
-python3 qso_hour_band.py
+python3 ../lib/qso_hour_band.py