+++ scripts/qso_months.sh
@@ -1,4 +1,4 @@
 DIR=`realpath -s $0|sed 's|\(.*\)/.*|\1|'`  
 echo Absolute path: ${DIR}
 cd ${DIR} 
-python3 qso_months.py
+python3 ../lib/qso_months.py