+++ scripts/qso_trend.sh
@@ -1,5 +1,5 @@
 DIR=`realpath -s $0|sed 's|\(.*\)/.*|\1|'`  
 echo Absolute path: ${DIR}
 cd ${DIR} 
-python3 qso_trend.py
+python3 ../lib/qso_trend.py