+++ static/html/replace_version.sh
@@ -1,0 +1,1 @@
+../../templates/replace_version.sh