+++ wsgi.py
@@ -1,6 +1,5 @@
-
 import bjoern
-from webapp import app
+from lib\webapp import app
 
 
 bjoern.listen(app, '0.0.0.0', 8080)